feat: implementado o teste referente ao exercício 6 e ajustado o tipo de retorno, pois o assert do junit não aceita integer por redundancia.

+++ testesUnitarios/src/main/java/easy/Exercicio05_easy.java
@@ -4,23 +4,23 @@ import javax.swing.*;
 
 public class Exercicio05_easy {//Inicio
 
-    public Integer calcularSoma(int primeiroValor, int segundoValor, int terceiroValor) { // inicio main
-        int soma = primeiroValor + segundoValor + terceiroValor; // calculo da soma
+    public long calcularSoma(long primeiroValor, long segundoValor, long terceiroValor) { // inicio main
+        long soma = primeiroValor + segundoValor + terceiroValor; // calculo da soma
         return soma;
     }// fim soma
 
-    public Integer calcularSubtracao(int primeiroValor, int segundoValor, int terceiroValor) {
-        int subtracao = segundoValor - primeiroValor; // Calculo subtracao
+    public long calcularSubtracao(long primeiroValor, long segundoValor, long terceiroValor) {
+        long subtracao = segundoValor - primeiroValor; // Calculo subtracao
         return subtracao;
     }
 
-    public Integer calcularMultiplicacao(int primeiroValor, int segundoValor, int terceiroValor) {
-        int multiplicacao = primeiroValor * segundoValor * terceiroValor;// Calculo multiplicação
+    public long calcularMultiplicacao(long primeiroValor, long segundoValor, long terceiroValor) {
+        long multiplicacao = primeiroValor * segundoValor * terceiroValor;// Calculo multiplicação
         return multiplicacao;
     }
 
-    public Integer calcularMedia(int primeiroValor, int segundoValor, int terceiroValor) {
-        int media = calcularMultiplicacao(primeiroValor, segundoValor, terceiroValor) / 3; // Calculo media
+    public long calcularMedia(long primeiroValor, long segundoValor, long terceiroValor) {
+        long media = calcularMultiplicacao(primeiroValor, segundoValor, terceiroValor) / 3; // Calculo media
         return media;
     }
